// src/types/property.ts
export interface ListingDetails {
  listingid: string;
  detailreq: string;
  location: string;
  locationstring: string;
  streetaddress: string;
  type: string;
  contract: string;
  beds: string;
  baths: string;
  isfurnished: string;
  garages: string;
  floorarea: string;
  title: string;
  description: string;
  price: string;
  pdr: string;
  leaseunit: string;
  leaseoptions: string[];
  isnegotiable: string;
  datelisted: string;
  imagelist: string[];
  amenities: string[];
  categorylink: string;
  categorytext: string;
  parentlink: string;
  parenttext: string;
  similars: ListingDetails[];
  image: string;
  owner: {
    haswan: boolean;
    name: string;
    page: string;
    type: string;
    accounttype: string;
    logo: string;
    profilepic: string;
    commission: string;
    registrationfee: string;
    listingscount: string;
    verification: string;
  };
}

export interface Unit {
  unitid: number;
  terms: string;
  unittype: string;
  beds: number;
  baths: number;
  garages: number;
  hasbalcony: number;
  price: number;
  floorarea: number;
  description: string;
  imcount: number | null;
  projectid: number;
  coverphoto: string;
  dateadded: string;
  soldout: number;
  csign: string;
  unitname: string;
  fullyfurnished: number;
  availableunits: number;
  featured: number;
  test: number;
  developerid: number;
  rentduration: number;
  parkingspace: number;
  sellingpricecsign: string;
  rentpricecsignpermonth: string;
  rentpricecsignperweek: string;
  rentpricecsignperday: string;
  rentdurationindays: number;
  rentdurationinweeks: number;
  rentdurationinmonths: number;
  rentpricepermonth: number;
  rentpriceperweek: number;
  rentpriceperday: number;
  sellingprice: number;
  rentdurationtype: string;
  hasparkingspace: number;
  city: string;
  address: string;
  tourvideo: string;
  publish: number;
  lng: number;
  lat: number;
  formatted_address: string;
  pageviews: number;
  negotiable: number;
  landmesurement: string;
  landlength: number;
  landwidth: number;
  updated_at: string;
  title: string;
  persquaremeter: number;
  photostorage: string;
  unittypename: string;
  unittypeslug: string;
  contactclicks: number;
  topad: string | null;
}

export interface FloorPlan {
  id: string;
  sqft: number;
  sqm: number;
  beds: number;
  baths: number;
  type: string;
  title: string;
  imageUrl?: string;
  unitNumbers?: string;
}